+++ src/context/GamepadContext.ts
@@ -1,10 +1,10 @@
 import React from "react";
-import { GamepadButtonName } from "../types";
+import { GamepadAxisName, GamepadButtonName } from "../types";
 
 type GamepadContextProps = {
   isButtonDown: (index: number, button: GamepadButtonName) => boolean;
   getButtonAxis: (index: number, negative: GamepadButtonName, positive: GamepadButtonName) => number;
-  getAnalogAxis: (axis: number) => number;
+  getAnalogAxis: (index:number, axis: GamepadAxisName) => number;
 }
 
 export const GamepadContext = React.createContext<GamepadContextProps>({