+++ src/hooks/useGamepadButtonMap.ts
@@ -1,15 +1,14 @@
-import { GamepadButtonName, Prop } from "../types";
+import { GamepadButtonMap, GamepadButtonName, Prop } from "../types";
 import { useGamepad } from "./useGamepad";
 import { useProperty } from "./useProperty";
 import { useUpdate } from "./useUpdate";
 import { useVirtualInput } from "./useVirtualInput";
 
-type GamepadMap = Partial<Record<GamepadButtonName, string>>;
-
-export const useGamepadMap = (index: number, map: GamepadMap, active?: Prop<boolean>) => {
+export const useGamepadButtonMap = (index: Prop<number>, map: GamepadButtonMap, active?: Prop<boolean>) => {
   const { simulate } = useVirtualInput();
   const { isButtonDown } = useGamepad();
 
+  const gamepadIndex = useProperty(index);
   const isActive = useProperty(active === undefined ? true : active);
 
   useUpdate(() => {
@@ -17,7 +16,7 @@ export const useGamepadMap = (index: number, map: GamepadMap, active?: Prop<bool
       for (const button in map) {
         const action = map[button as GamepadButtonName];
         
-        if (action && isButtonDown(index, button as GamepadButtonName)) {
+        if (action && isButtonDown(gamepadIndex.current, button as GamepadButtonName)) {
           simulate(action);
         }
       }